+++ src/main/java/com/sim/wifi/upgrade/modules/permissions/service/AbUserCacheService.java
@@ -36,7 +36,7 @@ public interface AbUserCacheService {
     /**
      * 当资源信息改变时，删除资源项目后台用户缓存
      */
-    void delResourceListByResource(Long resourceId);
+    void delResourceListByResource(Integer resourceId);
 
     /**
      * 获取缓存后台用户信息
